00_startcamp/03_AI/vibe_coding/client_teacher/requirements.txt
PyQt5
websocket-client
PyInstaller
